separate kysely types to separate file from schema
since this will introduce a breaking change, bumping version to 0.4.0.
BREAKING CHANGES:
  * anywhere we are importing from `db/schema` will likely need to
    change to `db/types`
