+++ spec/unit/helpers/cli/SchemaBuilder.spec.ts
@@ -3,7 +3,7 @@
 // written
 
 import User from '../../../../test-app/app/models/User'
-import { BalloonColorsEnumValues } from '../../../../test-app/db/schema'
+import { BalloonColorsEnumValues } from '../../../../test-app/db/types'
 
 describe('SchemaBuilder', () => {
   describe('#build', () => {
