+++ src/helpers/path/dbTypesPath.ts
@@ -1,0 +1,8 @@
+import path from 'path'
+import loadDreamYamlFile from './loadDreamYamlFile'
+import projectRootPath from './projectRootPath'
+
+export default async function dbTypesPath({ omitDirname }: { omitDirname?: boolean } = {}) {
+  const yamlConfig = await loadDreamYamlFile()
+  return projectRootPath({ filepath: path.join(yamlConfig.db_path, 'types.ts'), omitDirname })
+}
