+++ test-app/app/models/Balloon.ts
@@ -3,7 +3,7 @@ import { IdType } from '../../../src/dream/types'
 import HasOne from '../../../src/decorators/associations/has-one'
 import Scope from '../../../src/decorators/scope'
 import Validates from '../../../src/decorators/validations/validates'
-import { BalloonColorsEnum, BalloonTypesEnum } from '../../db/schema'
+import { BalloonColorsEnum, BalloonTypesEnum } from '../../db/types'
 import BalloonLine from './BalloonLine'
 import ApplicationModel from './ApplicationModel'
 import Sortable from '../../../src/decorators/sortable'
